+++ changelog.md
@@ -1,10 +1,31 @@
 # Changelog
 
-**./test-parts/test-submission.js**
-* New file - Unit testing for the assessment submission itself.
-* Placeholder only.
+**./cases.js**
+* New file
+	* Test cases will be performed here rather than in the main file.
+	* Equally accessible between main submission and unit tests.
+	* Returns both expected and actual output.
 
-**./test/index.js**
-* Added requirement: '../test-parts/test-submission'
-* Commented out 'testPathfindingInput' call.
-* Call to 'testSubmission'
+---
+
+**./submission.js**
+* Added new requirement: './cases'
+* Removed requirements:
+	* exactRoute
+	* shortestRoute
+	* possibleRoutes
+	* routeCriteria
+	* numSigns
+* Removed globals:
+	* routeStopCriteria
+	* routeDistCriteria
+* Wrote new function 'callTestCases'. Merges the following:
+	* callExactRouteTestCases
+	* callPossibleRouteByStopCountTestCases
+	* callShortestRouteTestCases
+	* callPossibleRouteByDistanceTestCase
+
+---
+
+**./src/result-display.js**
+* Revised to use new test case structure.